python/requirements.txt
alembic==1.13.1
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
google-cloud-pubsub==2.21.5
apache-beam==2.57.0
apache-beam[gcp,test,docs]==2.57.0
